AttributeSideBarService listens to selectedBuilding #527
