+++ visualization/app/codeCharta/ui/attributeSideBar/attributeSideBar.component.html
@@ -1,6 +1,6 @@
 <div class="side-bar-container cc-shadow">
 	<div class="side-bar-header">
-		<i class="close-icon fa fa-times" ng-click="$ctrl.closeSideBar()"></i>
+		<i class="close-icon fa fa-times" ng-click="$ctrl.onClickCloseSideBar()"></i>
 
 		<a ng-show="$ctrl._viewModel.node.link" class="node-link" href="{{ $ctrl._viewModel.node.link }}" target="_blank">
 			<i class="fa fa-external-link"></i>
