+++ visualization/app/codeCharta/ui/attributeSideBar/attributeSideBar.component.ts
@@ -1,11 +1,7 @@
 import "./attributeSideBar.component.scss"
-import { IRootScopeService, ITimeoutService } from "angular"
+import { IRootScopeService } from "angular"
 import { CodeMapBuilding } from "../codeMap/rendering/codeMapBuilding"
-import {
-	CodeMapMouseEventService,
-	BuildingSelectedEventSubscriber,
-	BuildingDeselectedEventSubscriber
-} from "../codeMap/codeMap.mouseEvent.service"
+import { CodeMapMouseEventService, BuildingSelectedEventSubscriber } from "../codeMap/codeMap.mouseEvent.service"
 import { Node } from "../../codeCharta.model"
 import _ from "lodash"
 import {
@@ -15,7 +11,7 @@ import {
 	EdgeMetricSubscriber
 } from "../../state/settingsService/settings.service.events"
 import { SettingsService } from "../../state/settingsService/settings.service"
-import { AttributeSideBarService } from "./attributeSideBar.service"
+import { AttributeSideBarService, AttributeSideBarVisibilitySubscriber } from "./attributeSideBar.service"
 import $ from "jquery"
 
 interface PrimaryMetrics {
@@ -28,11 +24,11 @@ interface PrimaryMetrics {
 export class AttributeSideBarController
 	implements
 		BuildingSelectedEventSubscriber,
-		BuildingDeselectedEventSubscriber,
 		AreaMetricSubscriber,
 		HeightMetricSubscriber,
 		ColorMetricSubscriber,
-		EdgeMetricSubscriber {
+		EdgeMetricSubscriber,
+		AttributeSideBarVisibilitySubscriber {
 	private _viewModel: {
 		node: Node
 		primaryMetricKeys: PrimaryMetrics
@@ -44,29 +40,18 @@ export class AttributeSideBarController
 	}
 
 	/* @ngInject */
-	constructor(
-		private $rootScope: IRootScopeService,
-		private $timeout: ITimeoutService,
-		private attributeSideBarService: AttributeSideBarService
-	) {
+	constructor(private $rootScope: IRootScopeService, private attributeSideBarService: AttributeSideBarService) {
 		CodeMapMouseEventService.subscribeToBuildingSelectedEvents(this.$rootScope, this)
-		CodeMapMouseEventService.subscribeToBuildingDeselectedEvents(this.$rootScope, this)
 		SettingsService.subscribeToAreaMetric(this.$rootScope, this)
 		SettingsService.subscribeToHeightMetric(this.$rootScope, this)
 		SettingsService.subscribeToColorMetric(this.$rootScope, this)
 		SettingsService.subscribeToEdgeMetric(this.$rootScope, this)
+		AttributeSideBarService.subscribe(this.$rootScope, this)
 	}
 
 	public onBuildingSelected(selectedBuilding: CodeMapBuilding) {
 		this._viewModel.node = selectedBuilding.node
 		this.updateSortedMetricKeysWithoutPrimaryMetrics()
-		this.openSideBar()
-		this.synchronizeAngularTwoWayBinding()
-	}
-
-	public onBuildingDeselected() {
-		this.closeSideBar()
-		this.synchronizeAngularTwoWayBinding()
 	}
 
 	public onAreaMetricChanged(areaMetric: string) {
@@ -89,14 +74,16 @@ export class AttributeSideBarController
 		this.updateSortedMetricKeysWithoutPrimaryMetrics()
 	}
 
-	public closeSideBar() {
-		$(".side-bar-container").removeClass("expanded")
-		this.attributeSideBarService.closeSideBar()
+	public onAttributeSideBarVisibilityChanged(isAttributeSideBarVisible: boolean) {
+		if (isAttributeSideBarVisible) {
+			$(".side-bar-container").addClass("expanded")
+		} else {
+			$(".side-bar-container").removeClass("expanded")
+		}
 	}
 
-	public openSideBar() {
-		$(".side-bar-container").addClass("expanded")
-		this.attributeSideBarService.openSideBar()
+	public onClickCloseSideBar() {
+		this.attributeSideBarService.closeSideBar()
 	}
 
 	private updateSortedMetricKeysWithoutPrimaryMetrics() {
@@ -105,10 +92,6 @@ export class AttributeSideBarController
 				.filter(x => !_.values(this._viewModel.primaryMetricKeys).includes(x))
 				.sort()
 		}
-	}
-
-	private synchronizeAngularTwoWayBinding() {
-		this.$timeout(() => {})
 	}
 }
